ST-23: Integrate student views for all existing pages
- Adjust student view for dashboard page and all classroom related
  pages.
- Refactor Dashboard component structure.
- Add CardsContainer, ClassroomCard and Post components.
- Add (you) tag when currentUser encounter their name/posts.
