+++ src/Pages/Dashboard/Components/CreateClassroomButton/CreateClassroomButton.jsx
@@ -11,6 +11,7 @@ const CreateClassroomButton = () => {
   const closeClassroomFormModal = () => {
     setIsClassroomFormModalOpen(false);
   };
+
   return (
     <>
       <Button
