+++ src/Pages/People/Components/StudentHeadingAndAddButton/StudentHeadingAndAddButton.jsx
@@ -1,17 +1,22 @@
 import { ActionIcon, Flex, Title } from "@mantine/core";
 import { IconSquareRoundedPlus } from "@tabler/icons-react";
 
-const StudentHeadingAndAddButton = ({ setIsAddStudentModalOpen }) => {
+const StudentHeadingAndAddButton = ({
+  setIsAddStudentModalOpen,
+  currentUser,
+}) => {
   return (
     <Flex align={"center"} justify={"space-between"} mt={"xl"}>
       <Title order={2}>Students</Title>
-      <ActionIcon
-        variant="subtle"
-        color="sazim-green"
-        onClick={() => setIsAddStudentModalOpen(true)}
-      >
-        <IconSquareRoundedPlus />
-      </ActionIcon>
+      {currentUser.role === "teacher" ? (
+        <ActionIcon
+          variant="subtle"
+          color="sazim-green"
+          onClick={() => setIsAddStudentModalOpen(true)}
+        >
+          <IconSquareRoundedPlus />
+        </ActionIcon>
+      ) : null}
     </Flex>
   );
 };
